create styles/style_consts.dart to store consts, access theme from context in course_detail_widgets
