+++ lib/widgets/course_detail_widgets.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:we_find/styles/theme.dart' as Theme;
 
 class GroupDetails extends StatelessWidget {
   /* TODO: Uncomment this once this class exists
@@ -12,6 +11,7 @@ class GroupDetails extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final ThemeData themeData = Theme.of(context);
     return Container(
       margin: const EdgeInsets.all(15),
       padding: const EdgeInsets.all(15),
@@ -20,20 +20,20 @@ class GroupDetails extends StatelessWidget {
           // subtitle
           Text(
             "Max participants:" + ' ' + "42",
-            style: Theme.Text.heading3Style,
+            style: themeData.textTheme.headline3,
           ),
           // build lecturers list
           //_buildLecturersList(Lecturers[] lecturers),
           // subtitle
           Text(
             "Group information:",
-            style: Theme.Text.heading3Style,
+            style: themeData.textTheme.headline3,
           ),
           // regular text
           SelectableText(
             "Sample Course info1234567897 alalallalalalal." +
                 "Should be able to copy from here!",
-            style: Theme.Text.textStyle,
+            style: themeData.textTheme.bodyText1,
           ),
           //_buildTimeTable(TimeTableEntry[] timeTable)
         ],
